+++ src/FileUpload/FileUpload.tsx
@@ -1,203 +1,12 @@
 import { useState, useRef, useMemo, forwardRef, useEffect } from 'react'
-import styled, { css, keyframes } from 'styled-components'
+import * as Styled from './FileUpload.styled'
 import { Button } from '../Button'
 import { Icon, IconType } from '../Icon'
 import { FileCard } from '../FileCard'
 import { Spacer } from '../Layout/Spacer'
 import { SaveButton } from '../SaveButton'
 import { AcceptType } from './fileTypes'
-
-const UploadForm = styled.form`
-  min-height: 160px;
-  min-width: 300px;
-  height: 100%;
-  width: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-
-  .header {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    h2 {
-      margin: 0;
-    }
-  }
-
-  .upload-button {
-    position: relative;
-    overflow: hidden;
-
-    * > span {
-      cursor: pointer;
-    }
-  }
-
-  input {
-    position: absolute;
-    inset: 0px;
-    opacity: 0;
-
-    &::-webkit-file-upload-button {
-      visibility: hidden;
-    }
-
-    &::after {
-      content: '';
-      position: absolute;
-      inset: 0px;
-      cursor: pointer;
-    }
-  }
-
-  .files {
-    flex: 1;
-    position: relative;
-    background-color: var(--panel-background);
-    border-radius: 1rem;
-
-    border-width: 2px;
-    border-style: dashed;
-    border-color: #6b7685;
-  }
-
-  .scroll-container {
-    overflow: auto;
-    position: absolute;
-    inset: 0;
-  }
-
-  label {
-    position: absolute;
-    inset: 0;
-
-    &.drag-active {
-      background-color: rgba(0, 0, 0, 0.1);
-      z-index: 100;
-    }
-  }
-
-  small {
-    color: #e53e3e !important;
-    font-size: 0.8rem;
-  }
-
-  .drop-here {
-    position: absolute;
-    inset: 0;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    user-select: none;
-
-    .icon {
-      font-size: 3rem;
-    }
-  }
-
-  #drag-file-element {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    border-radius: 1rem;
-    inset: 0;
-    z-index: 25;
-  }
-
-  footer {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-
-    .success,
-    .error {
-      display: block;
-      &::after {
-        content: ':';
-        opacity: 0;
-      }
-    }
-
-    .success {
-      color: var(--color-hl-00);
-    }
-
-    .error {
-      color: var(--color-hl-error);
-    }
-  }
-`
-
-// fade out and scale down
-const successAnimation = keyframes`
-  from {
-    opacity: 1;
-    transform: translateY(0);
-  }
-  to {
-    opacity: 0;
-    transform: translateY(-50px);
-  }
-`
-
-const StyledList = styled.ul<{ $isSuccess: boolean }>`
-  list-style: none;
-  padding: 16px;
-  margin: 0;
-
-  position: relative;
-  z-index: 50;
-
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-
-  overflow: auto;
-
-  ${({ $isSuccess }) =>
-    $isSuccess &&
-    css`
-      animation: ${successAnimation} 0.3s ease forwards;
-    `}
-`
-
-const extractSequence = (string: string): [string, number] | [] => {
-  // first remove the extension
-
-  const matches = string.match(/\d+/g) // Extracts all sequences of digits from the string
-  if (!matches) return []
-
-  const lastMatch = matches[matches.length - 1] // Gets the last match, which is the sequence number
-  const prefix = string.slice(0, string.lastIndexOf(lastMatch)) // Gets the prefix of the string, which is the part before the sequence number
-
-  const sequenceNumber = parseInt(lastMatch, 10) // Parses the sequence number into an integer
-  return [prefix, sequenceNumber]
-}
-
-const getSeqError = (files: CustomFile[]) => {
-  // message example "Frames failed to upload : 0003, 0004"
-  const prefix = 'Errors on frames : '
-  const failedFrames = []
-
-  // console.log(files)
-
-  for (const file of files) {
-    if (!file.message) continue
-
-    failedFrames.push(file.sequenceNumber)
-  }
-
-  if (!failedFrames.length) return undefined
-
-  const message = prefix + failedFrames.join(', ')
-
-  return message
-}
+import { extractSequence, getSeqError } from './FileUploadHelpers'
 
 export interface CustomFile {
   sequenceId: string | null
@@ -621,7 +430,7 @@ export const FileUpload = forwardRef<HTMLFormElement, FileUploadProps>(
     }${accept.length ? ' ' + accept.join(', ') : ' All Files Types'}`
 
     return (
-      <UploadForm
+      <Styled.Form
         onDragEnter={handleDrag}
         onSubmit={(e) => e.preventDefault()}
         ref={ref}
@@ -672,7 +481,7 @@ export const FileUpload = forwardRef<HTMLFormElement, FileUploadProps>(
           />
 
           <div className="scroll-container">
-            <StyledList
+            <Styled.List
               $isSuccess={!!isSuccess && !!localFilesState.length}
               onAnimationEnd={() => setLocalFilesState([])}
               style={listStyle}
@@ -695,7 +504,7 @@ export const FileUpload = forwardRef<HTMLFormElement, FileUploadProps>(
                   />
                 </li>
               ))}
-            </StyledList>
+            </Styled.List>
           </div>
 
           {!filesToGroup.length && (
@@ -748,7 +557,7 @@ export const FileUpload = forwardRef<HTMLFormElement, FileUploadProps>(
             )}
           </footer>
         )}
-      </UploadForm>
+      </Styled.Form>
     )
   },
 )
